@@ -4,18 +4,19 @@
   "private": true,
   "dependencies": {
     "@babel/runtime": "^7.27.6",
-    "@material-ui/core": "^4.7.2",
-    "@material-ui/icons": "^4.5.1",
+    "@mui/material": "^5.15.0",
+    "@mui/icons-material": "^5.15.0",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@types/jest": "^24.0.23",
     "@types/lodash": "^4.14.149",
     "@types/node": "^12.12.16",
-    "@types/react": "^16.9.16",
-    "@types/react-dom": "^16.9.4",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "axios": "^0.30.0",
-    "bootstrap": "^5.0.1",
     "lodash": "4.17.21",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "typescript": "^3.7.3"
   },
