@@ -1,7 +1,7 @@
 import React, { useReducer, createContext } from 'react';
-import { Grid } from '@material-ui/core';
+import { Grid, Typography, Box } from '@mui/material';
 import InterestifyTable from '../components/FunctionalTable';
-import ApiService  from '../services/ApiService';
+import ApiService from '../services/ApiService';
 import Applet from '../components/Applet';
 import { searchReducer, SearchState } from '../common/Reducers';
 import SearchForm from '../components/SearchForm';
@@ -30,9 +30,9 @@ const SearchFollowers = (props: SearchFollowersProps) => {
         error: '',
         searchKey: ''
     };
-    
-    const [ state, dispatch ] = useReducer(searchReducer, initialState);
-    
+
+    const [state, dispatch] = useReducer(searchReducer, initialState);
+
     const StateContext = createContext(state); // context containing state
 
     // search logic
@@ -55,24 +55,39 @@ const SearchFollowers = (props: SearchFollowersProps) => {
     };
 
     return (
-        <Applet title="Search Followers" id={ id } width={ width } height={ height } StateContext={ StateContext } run={ run }>
-            <Grid container spacing={ 2 }>
-                <Grid item xs={ 12 }>
-                    <SearchForm searchKey={ state.searchKey } dispatch={ dispatch } />
+        <Applet title="Search Followers" id={id} width={width} height={height} StateContext={StateContext} run={run}>
+            <Grid container spacing={3}>
+                <Grid item xs={12}>
+                    <SearchForm searchKey={state.searchKey} dispatch={dispatch} />
                 </Grid>
-                <Grid className="container" item xs={ 12 }>
+                <Grid item xs={12}>
                     {
                         state.tableData.body.length > 0 ? 
-                        <InterestifyTable header={ state.tableData.header } body={ state.tableData.body } striped sortable missingValue="N/A"
-                            pagination={ {
-                            rowsPerPageOptions: [ 5, 10, 15 ],
-                            rowsPerPage: 5,
-                            } }
-                            sort={ {
-                            id: 'retweetCount',
-                            order: 'desc'
-                            } } /> :
-                        <div></div>
+                        <InterestifyTable 
+                            header={state.tableData.header} 
+                            body={state.tableData.body} 
+                            striped 
+                            sortable 
+                            missingValue="N/A"
+                            pagination={{
+                                rowsPerPageOptions: [5, 10, 15],
+                                rowsPerPage: 5,
+                            }}
+                            sort={{
+                                id: 'retweetCount',
+                                order: 'desc'
+                            }} 
+                        /> :
+                        <Box 
+                            display="flex" 
+                            justifyContent="center" 
+                            alignItems="center" 
+                            minHeight="200px"
+                        >
+                            <Typography variant="body2" color="text.secondary">
+                                Enter a username to find followers
+                            </Typography>
+                        </Box>
                     }
                 </Grid>
             </Grid>
